Change Plan's property Presupuesto type for Decimal instead of double.

+++ BizData/Entities/Plan.cs
@@ -7,7 +7,7 @@ namespace BizData.Entities
     public class Plan
     {
         public int PlanID { get; set; }
-        public double Presupuesto { get; set; }
+        public Decimal Presupuesto { get; set; }
         public int Año { get; set; }
         public string TipoPlan { get; set; }
